perf: speed up mul, squaring, and Newton division
Four-phase GPU multi-precision optimization of TabaiInt:

- Phase 1 (micro-opts): lazy/propagated zero flags to skip syncing
  single-limb reads; add_trimmed/sub_trimmed/_trim_z return zero-ness
  for free; alias magnitude arrays instead of copying in __neg__/__abs__
  (arrays are immutable post-construction); drop redundant .astype in
  gpu_to_int; length-based _mag_cmp shortcut at the TabaiInt layer.
- Phase 2/3: faster schoolbook mul and squaring paths.
- Phase 4 (Newton division): size-gated ramp-width normalization
  (_NEWTON_RAMP_TRIM_LIMBS = 40000) — small widths use a deterministic
  sync-free top slice, large widths _trim to avoid a stray zero limb
  crossing an FFT power-of-two length and doubling the big mul; fixed
  single-step finish (was an up-to-6 loop with per-iter D2H compare);
  +b add-based step-4 adjustment; re-swept _DIV_NEWTON_THRESHOLD_LIMBS
  12288 -> 5120 after the faster mul moved the cpu//<->Newton crossover.

floordiv/mod @>=1e6 bit: 10-13% reduction; no regression at any size.
Tests: 482 passed (was 475; +7 incl. both-branch ramp-gate coverage);
test mirror constants updated in lockstep.

+++ tabai/src/tabai_gpu/utils.py
@@ -16,5 +16,7 @@ def int_to_gpu(n: int) -> cp.ndarray:
 
 
 def gpu_to_int(gpu_arr: cp.ndarray) -> int:
-    cpu_arr = cp.asnumpy(gpu_arr).astype(np.uint32)
+    # gpu_arr is always a uint32 little-endian limb array, so asnumpy already
+    # yields uint32 — the extra .astype(np.uint32) copy was redundant.
+    cpu_arr = cp.asnumpy(gpu_arr)
     return int.from_bytes(cpu_arr.tobytes(), byteorder='little')
